refactor(validations): convert files to typescript
