--- src/__internal__/validations/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from "./validation-icon.component";
-export { ValidationPropTypes } from "./validations";
