--- src/__internal__/validations/index.js
@@ -1,1 +1,0 @@
-export { default } from "./validation-icon.component";
