+++ src/__internal__/validations/index.ts
@@ -1,0 +1,5 @@
+export { default } from "./validation-icon.component";
+export type {
+  ValidationProps,
+  ValidationIconProps,
+} from "./validation-icon.component";
