+++ src/__internal__/validations/validation-icon.style.ts
@@ -2,15 +2,19 @@ import styled, { css } from "styled-components";
 import { margin } from "styled-system";
 import BaseTheme from "../../style/themes/base";
 import StyledIcon from "../../components/icon/icon.style";
+import { ValidationIconProps } from "./validation-icon.component";
 
 const validationIconTypes = {
   error: "var(--colorsSemanticNegative500)",
   info: "var(--colorsSemanticInfo500)",
-  success: "var(--colorsSemanticPositive500)",
   warning: "var(--colorsSemanticCaution500)",
 };
 
-const ValidationIconStyle = styled.span`
+type ValidationType = "error" | "warning" | "info";
+
+const ValidationIconStyle = styled.span<
+  ValidationIconProps & { validationType: ValidationType }
+>`
   background: none;
   cursor: default;
   display: flex;
